Rename package to work_model
