+++ sbt_common/work_model/src/main/scala/uk/ac/wellcome/work_model/Agent.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 sealed trait AbstractAgent {
   val label: String
