+++ sbt_common/work_model/src/main/scala/uk/ac/wellcome/work_model/Concept.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 sealed trait AbstractConcept {
   val label: String
