+++ sbt_common/work_model/src/main/scala/uk/ac/wellcome/work_model/License.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 import io.circe.{Decoder, Encoder, Json}
 import cats.syntax.either._
