+++ sbt_common/work_model/src/main/scala/uk/ac/wellcome/work_model/Place.scala
@@ -1,3 +1,3 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 case class Place(label: String, ontologyType: String = "Place")
