+++ sbt_common/work_model/src/test/scala/uk/ac/wellcome/work_model/IdentifierSchemesTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 class IdentifierSchemesTest extends FunSpec with Matchers with JsonTestUtil {
