+++ sbt_common/work_model/src/test/scala/uk/ac/wellcome/work_model/LicenseTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.work_model
 
 class LicenseTest extends FunSpec with Matchers {
